make `fold.expresion.cpp` work with negative values (#268)

+++ exercises/2/fold.expresion.cpp
@@ -11,7 +11,7 @@
 #include <iostream>
 template<typename ... T>
 auto average(T ... t) {
-    return (t + ... ) / sizeof...(t);
+    return static_cast<double>((t + ... )) / sizeof...(t);
 }
 int main() {
     std::cout << average(1, 2, 3, 4, 5, 6, 7, 8, 9, 10) << std::endl;
